fått ut product och user i page
