+++ prisma/schema.prisma
@@ -22,3 +22,8 @@ model Product {
   price       Int
   description String
 }
+
+model Category {
+  id   Int    @id @default(autoincrement())
+  name String @unique
+}
